requirements.txt
click==5.1
mecab-python3==0.7
numpy==1.10.1
scikit-learn==0.16.1
scipy==0.16.0
sklearn==0.0
